Refactor provider tests into dedicated modules
