+++ projects/04-llm-adapter-shadow/tests/providers/conftest.py
@@ -1,0 +1,8 @@
+from __future__ import annotations
+
+import pytest
+
+
+@pytest.fixture(scope="module")
+def provider_request_model() -> str:
+    return "gemini:test-model"
